fixed the fpga logic
we are now back to a single working LED using spi communication. There was an odd bug with the fpga that ended up causing the orientation to be off by one clk cycle. that is now fixed with a comment so that we can go back to it if it causes errors again in the future.